python/evaluation/requirements.txt
matplotlib
numpy
plotly
open3d
pandas
scipy
trimesh
GitPython
moviepy
kaleido